Keep compiled binary E2E coverage easier to extend
Queue item 103 targets the compiled binary integration suite, so this extracts the shared binary/server/browser helpers into a sibling test helper module while leaving the kitchen-sink assertions intact.

Constraint: Preserve compiled-binary E2E behavior and keep the refactor limited to test helper extraction in veryfront-code
Rejected: Split the giant test file into multiple scenario files | larger movement and review surface than this queue item needs
Confidence: high
Scope-risk: narrow
Directive: Keep heavy compiled-binary test harness helpers in the local integration test area unless another binary E2E suite needs the exact same lifecycle utilities
Tested: deno fmt --check tests/integration/compiled-binary-e2e.test.ts tests/integration/compiled-binary-e2e.test-helpers.ts; deno lint tests/integration/compiled-binary-e2e.test.ts tests/integration/compiled-binary-e2e.test-helpers.ts; deno check tests/integration/compiled-binary-e2e.test.ts tests/integration/compiled-binary-e2e.test-helpers.ts; VERYFRONT_BINARY=/tmp/veryfront-e2e-bin-worker2-103 deno test --allow-all tests/integration/compiled-binary-e2e.test.ts; git diff --check
Not-tested: Additional veryfront-code suites beyond the focused compiled-binary E2E coverage
